Moved CCXTOrder class and fix cancel method.
